requirements.txt
Flask
azure-storage-queue